+++ common/templates/yandex-metrika-inline.d.ts
@@ -1,0 +1,7 @@
+/// <reference path="../../types/index.d.ts" />
+
+declare function renderYandexMetrikaInline(options?: YandexMetrikaInlineOptions): string;
+declare function renderYandexMetrikaInlineHead(options?: YandexMetrikaInlineOptions): string;
+declare function renderYandexMetrikaInlineScript(options?: YandexMetrikaInlineOptions): string;
+
+export { renderYandexMetrikaInline, renderYandexMetrikaInlineHead, renderYandexMetrikaInlineScript };